Update: Modularise BreadcrumbNode component
