+++ codecov.yml
@@ -4,4 +4,5 @@ coverage:
   range: 98...100
   ignore:
     - "src/components/Main.js"
+    - "src/components/**/*.spec.*"
     - "src/examples"
