+++ src/components/alexandria/BreadcrumbNode/styles.scss
@@ -1,4 +1,4 @@
-@import '../color';
+@import 'styles/color';
 
 .breadcrumbnode-component {
   &-link {
